backend/pyproject.toml
[project]
name = "mood-sensor-backend"
version = "1.0"
description = "This is my attempt for the SEM exam, i generated this with gh:sicksid/django-nextjs-boilerplate and cookiecutter"
authors = [
    { name = "abdul hamid", email = "abdulachik@icloud.com" },
]

dependencies = [
    "django[gis]>=4.1.5",
    "djangorestframework>=3.14.0",
    "django-filter>=22.1",
    "markdown>=3.4.1",
    "drf-extensions>=0.7.1",
    "djangorestframework-simplejwt>=5.2.2",
    "django-health-check>=3.17.0",
    "django-extensions>=3.2.1",
    "django-environ>=0.9.0",
    "gunicorn>=20.1.0",
    "django-userforeignkey>=0.5.0",
    "pyyaml>=6.0",
    "uritemplate>=4.1.1",
    "whitenoise>=6.2.0",
    "psycopg2-binary>=2.9.5",
    "django-cors-headers>=3.13.0",
    "djangorestframework-gis>=1.0",
]
requires-python = ">=3.11.1"
license = { text = "GPLv3" }

[build-system]
requires = ["pdm-pep517>=1.0.0"]
build-backend = "pdm.pep517.api"

[tool.pdm]
[tool.pdm.dev-dependencies]
dev = [
    "pytest>=7.2.1",
    "pytest-django>=4.5.2",
    "black[d]>=22.12.0",
    "pylint>=2.15.10",
    "model-bakery>=1.9.0",
    "pylint-django>=2.5.3",
    "faker>=15.3.4",
    "pytest-testdox>=3.0.1",
    "pytest-env>=0.8.1",
]

[[tool.pdm.autoexport]]
filename = "requirements.txt"
without-hashes = true

[tool.pylint.DESIGN]
max-parents = 15

[tool.pylint.FORMAT]
max-line-length = 120

[tool.pylint.MASTER]
init-hook = "import os, sys; sys.path.append(os.path.dirname('.'))"
load-plugins = "pylint_django"
django-settings-module = "api.settings"
fail-under = 5.00
ignore-patterns = ["__pypackages__",
    "fixtures",
    "docs",
    ".venv"]
disable = ["C0114", # missing-module-docstring
    "C0115", # missing-class-docstring
    "C0116", # missing-function-docstring
    "C0321", # multiple-statements
    "W0221", # arguments-differ
    "W0222", # signature-differs
    "E1131", # unsupported-binary-operation
    "R0903", ] # too-few-public-methods

[tool.black]
target-version = ['py310']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.pytest_cache
  | __pypackages__
  | .venv
  | migrations
  | settings
  | apps.py
  | urls.py
  | wsgi.py
  | asgi.py
  | manage.py
  | __version__.py
)/
'''


[tool.pdm.scripts]
_.env_file = ".env"
test = "pytest"
start = { cmd = "python manage.py runserver" }
format = "black ."
lint = "black . --check --diff"

